user_management/views.py
from django.shortcuts import render

# Create your views here.

from allauth.account.views import SignupView
from .forms import CustomSignupForm


class CustomSignupView(SignupView):
    form_class = CustomSignupForm